pyproject.toml
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
 name = "sneakyagent"
 version = "0.1.0"
 description = "Workspace-context red teaming tool for AI coding agents"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
   "typer>=0.9.0",
   "pyyaml>=6.0.1",
   "openai>=1.0.0",
 ]
 
 [project.scripts]
 sneakyagent = "sneakyagent.cli:app"
 
 [tool.hatch.build.targets.wheel]
 packages = ["sneakyagent"]
include = ["sneakyagent/data/*.yaml"]
